requirements.txt
# ---------------- Core Backend ----------------
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1

# ---------------- Database ----------------
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9          # PostgreSQL connector
pymongo==4.8.0                  # MongoDB client
redis==5.0.7                    # For Celery + caching

# ---------------- Task Queue ----------------
celery==5.4.0
kombu==5.4.0

# ---------------- AI / ML ----------------
google-generativeai==0.7.0      # Gemini API client
openai==1.47.0                  # Whisper / fallback LLMs
sentence-transformers==3.1.1    # Embeddings for RAG
faiss-cpu==1.8.0.post1          # Vector DB for retrieval
scikit-learn==1.5.1             # Clustering, LDA
bertopic==0.16.2                # Topic modeling for disinfo detection
transformers==4.44.2            # HuggingFace transformers
                   # Needed for embeddings + models

# ---------------- NLP Utils ----------------
nltk==3.9.1
spacy==3.8.2
textblob==0.18.0.post0

# ---------------- Social Media APIs ----------------
tweepy==4.14.0                  # Twitter/X API
praw==7.7.1                     # Reddit API

# ---------------- Media / Provenance ----------------
pillow==10.4.0                  # Image handling
opencv-python==4.10.0.84        # Video/image processing
exifread==3.0.0                 # EXIF metadata
pycose==1.0.0        # Handle C2PA COSE signatures
cbor2==5.6.3         # Needed for COSE/CBOR payloads

# ---------------- Web & Security ----------------
httpx==0.27.2                   # Async HTTP client
requests==2.32.3
cryptography==43.0.1            # Ledger hashes, JWT signing
passlib[bcrypt]==1.7.4          # Password hashing if auth needed
python-jose==3.3.0              # JWT tokens for RBAC

# ---------------- Observability ----------------
loguru==0.7.2
prometheus-client==0.21.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-prometheus==0.46b0


# ---------------- Testing ----------------
pytest==8.3.2
pytest-asyncio==0.23.8
httpx[cli]==0.27.2              # For testing APIs


# Deepfake Detection
deepface==0.0.95   # auto-uses PyTorch backend
torch==2.5.1
torchvision==0.20.1
transformers==4.44.2
timm==1.0.9
pillow==10.4.0
accelerate==0.34.2

# Metadata Validation
exifread==3.0.0

# Graph Analysis
networkx==3.3

# Model Hub
huggingface-hub==0.24.6
